moved the create user & report to appropriate folders
