+++ pages/api/report/create-report.js
@@ -1,4 +1,4 @@
-import prisma from '../../prisma/client'
+import prisma from '../../../prisma/client'
 
 //POST api/create-report
 export default async function handler(req, res) {
